+++ example/src/pages/UploadPage.tsx
@@ -20,7 +20,17 @@ const Example = () => {
         }}
       />
       <br />
-      <Upload uploadType="picture" accept="image/*" />
+
+      <Upload
+        uploadType="picture"
+        accept="image/*"
+        requestUpload={{
+          endpoint: "https://www.mocky.io/v2/5cc8019d300000980a055e76",
+          headers: {
+            authorization: "authorization-text"
+          }
+        }}
+      />
     </Container>
   );
 };